third_party/WebKit/LayoutTests/fast/viewport/viewport-133-expected.txt

CONSOLE MESSAGE: line 3: The value "10;" for key "minimum-scale" was truncated to its numeric prefix.
CONSOLE MESSAGE: line 3: Error parsing a meta element's content: ';' is not a valid key-value pair separator. Please use ',' instead.
ALERT: viewport size 980x1078 scale 10 with limits [10, 10] and userScalable true
